+++ auto_route/example/lib/mobile/screens/books/routes.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:auto_route/empty_router_widgets.dart';
 import 'package:example/mobile/router/auth_guard.dart';
-import 'package:example/mobile/screens/empty/empty_router_page.dart';
 
 import 'book_details_page.dart';
 import 'book_list_page.dart';