+++ auto_route/test/deferred_router/router.dart
@@ -20,6 +20,10 @@ class EmptyRouterPage extends AutoRouter {
   const EmptyRouterPage({Key? key}) : super(key: key);
 }
 
-class SecondNested1Page extends TestPage {}
+class SecondNested1Page extends TestPage {
+  const SecondNested1Page({Key? key}) : super(key: key);
+}
 
-class SecondNested2Page extends TestPage {}
+class SecondNested2Page extends TestPage {
+  const SecondNested2Page({Key? key}) : super(key: key);
+}