+++ auto_route/test/nested_router/router.dart
@@ -1,4 +1,5 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:auto_route/empty_router_widgets.dart';
 import 'package:flutter/material.dart';
 
 import '../test_page.dart';
@@ -20,9 +21,7 @@ part 'router.gr.dart';
 )
 class AppRouter extends _$AppRouter {}
 
-class EmptyRouterPage extends AutoRouter {
-  const EmptyRouterPage({Key? key}) : super(key: key);
-}
+
 
 class SecondNested1Page extends TestPage {
   const SecondNested1Page({Key? key}) : super(key: key);