+++ auto_route/test/nested_tabs_router/router.dart
@@ -1,7 +1,7 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:auto_route/empty_router_widgets.dart';
 import 'package:flutter/material.dart';
 
-import '../nested_router/router.dart';
 import '../test_page.dart';
 import 'router_test.dart';