+++ auto_route_generator/lib/src/code_builder/library_builder.dart
@@ -29,8 +29,11 @@ String generateLibrary(
 }) {
   final fileName = config.element.source.uri.pathSegments.last;
 
-  throwIf(usesPartBuilder && deferredLoading,
-      'Part-file approach will not work with deferred loading because allocator need to mark all deferred imports!');
+  throwIf(
+    usesPartBuilder && deferredLoading,
+    'Part-file approach will not work with deferred loading because allocator needs to mark all deferred imports!',
+    element: config.element,
+  );
 
   final emitter = DartEmitter(
     allocator: usesPartBuilder ? Allocator.none : DeferredPagesAllocator(config.routes, deferredLoading),
@@ -40,6 +43,19 @@ String generateLibrary(
 
   var allRouters = config.collectAllRoutersIncludingParent;
   List<RouteConfig> allRoutes = allRouters.fold(<RouteConfig>[], (acc, a) => acc..addAll(a.routes));
+
+  final deferredRoutes = allRoutes.where((r) => r.deferredLoading == true);
+  throwIf(
+    usesPartBuilder && deferredRoutes.isNotEmpty,
+    'Part-file approach will not work with deferred loading because allocator needs to mark all deferred imports! ${deferredRoutes.map((e) => e.name)}',
+  );
+
+  for(var i = 0; i < allRoutes.length; i++){
+    final route = allRoutes[i];
+     if(deferredRoutes.any((e) => e.pageType == route.pageType && route.deferredLoading != true)){
+        allRoutes[i] = route.copyWith(deferredLoading: true);
+     }
+  }
 
   final nonRedirectRoutes = allRoutes.where((r) => r.routeType != RouteType.redirect);
   final checkedRoutes = <RouteConfig>[];
@@ -63,7 +79,7 @@ String generateLibrary(
         if (usesPartBuilder) Directive.partOf(fileName),
       ])
       ..body.addAll([
-        buildRouterConfig(config, allGuards, allRoutes, deferredLoading),
+        buildRouterConfig(config, allGuards, allRoutes),
         ...allRoutes
             .where((r) => r.routeType != RouteType.redirect)
             .distinctBy((e) => e.routeName)