+++ auto_route_generator/lib/src/resolvers/route_config_resolver.dart
@@ -35,6 +35,8 @@ class RouteConfigResolver {
     var path = autoRoute.peek('path')?.stringValue;
     var isDeferred = autoRoute.peek('deferredLoading')?.boolValue ??
         _routerConfig.deferredLoading;
+
+
     if (page == null) {
       var redirectTo = autoRoute.peek('redirectTo')?.stringValue;
       throwIf(