+++ auto_route_generator/lib/src/resolvers/router_config_resolver.dart
@@ -97,6 +97,8 @@ class RouterConfigResolver {
     );
 
     var routes = _resolveRoutes(routerConfig, autoRoutes);
+
+
     return routerConfig.copyWith(routes: routes);
   }